+++ packages/package-json-lint/src/rules/index.js
@@ -1,11 +1,11 @@
 const requireThemeContextVersions = require('./require-theme-context-versions');
 const requireNoTerraBasePeerDependencyVersions = require('./require-no-terra-base-peer-dependency-versions');
-const requireNoUnnecessaryDependency = require('./require-no-unnecessary-dependency');
+const requireDependenciesDeclaredAtAppropriateLevel = require('./require-dependencies-declared-at-appropriate-level');
 const requireNoHardCodedDependencyVersions = require('./require-no-hard-coded-dependency-versions');
 
 module.exports = {
-  'require-no-terra-base-peer-dependency-versions': requireNoTerraBasePeerDependencyVersions,
-  'require-no-unnecessary-dependency': requireNoUnnecessaryDependency,
-  'require-theme-context-versions': requireThemeContextVersions,
+  'require-dependencies-declared-at-appropriate-level': requireDependenciesDeclaredAtAppropriateLevel,
   'require-no-hard-coded-dependency-versions': requireNoHardCodedDependencyVersions,
+  'require-no-terra-base-peer-dependency-versions': requireNoTerraBasePeerDependencyVersions,
+  'require-theme-context-versions': requireThemeContextVersions,
 };